requirements.txt
discord-webhook == 1.0.*
beautifulsoup4 == 4.11.*
feedparser == 6.0.*
python-dotenv == 0.21.*